refactoring of the code, broke stuff out of app.jsx
